templates appear to be working

+++ extras/speed.py
@@ -14,8 +14,8 @@ print("OpenMP available: {}".format(pygram11.OPENMP))
 
 ipython = get_ipython()
 
-x = np.random.randn(1000000)
-w = np.random.uniform(0.8, 1.2, len(x))
+x = np.random.randn(1000000).astype(np.float32)
+w = np.random.uniform(0.8, 1.2, len(x)).astype(np.float32)
 nbins = 20
 xmin = -3
 xmax = 3
